fix(test): make the stale sweep safe against a name it cannot parse
Three findings from the architecture review.

The sweep tested `datname ~ '^pma_test_[0-9]+_'` and cast `split_part(...)` to
bigint in the same WHERE clause. PostgreSQL orders qualifiers by cost, not
left to right, so a database called `pma_test_x_y` on the shared server could
have raised 22P02 and failed every integration run until someone renamed it.
One `substring(datname from $2)` does both jobs, and a non-match is NULL
rather than an error.

`sweepStaleClones` takes the name pattern, so its own test sweeps fixtures
under a prefix of its own. Creating a real `pma_test_` clone and then holding
it open left a window in which a sibling worktree's sweep could drop it — a
flake in the test that exists to prove parallel runs are safe.

`products.pricing_rules_version` becomes bigint: the value is the rule set's
max(updated_at) in epoch seconds, which leaves int4 in 2038.

ADR-0002 now states why age is the only available discriminator, why the
template is named after the checkout rather than per run, and that migration
0000 is regenerated in place only until the first deploy. README states the
one-run-per-worktree limit that the forced template rebuild implies.

+++ src/shared/db/migrations/0000_write_store.sql
@@ -59,7 +59,7 @@ CREATE TABLE "products" (
 	"category" text NOT NULL,
 	"base_price_cents" bigint NOT NULL,
 	"stock_quantity" integer NOT NULL,
-	"pricing_rules_version" integer,
+	"pricing_rules_version" bigint,
 	"ingest_job_id" bigint,
 	"ingest_source_offset" bigint,
 	"created_at" timestamp with time zone DEFAULT now() NOT NULL,
